Add a "label" to resolve collisions
The token secured and providing annotations may be used on the same method with the same type, but would need two distinct values. Create a "label" on each annotation and allow access in the ContextBean according to that label to set and get a value. Also retain the default functionality of not requiring a label to be specified.