Remove test app from the build until fixed
The updated version of Tycho has stricter version checks which fail when building the product. Attempting to remove the version constraints did not solve the issue. The build will likely need to depend on newer upstream sites before it will resolve, and the product build config will need more adjustment