Change LdapLoginProvider login user ID
The user ID was the LDAP DN, which changes based on the identity's OU hierarchy, and is expected to be mutable. The new user ID is based on the "objectGUID" field and has two values separated by a semicolon. The first value is the display value for the GUID with dash-separated groups, and the second value is the escaped byte sequence used to query LDAP